+++ src/features/expense/ui/SelectMonthDrawer/SelectMonthList.tsx
@@ -1,55 +1,6 @@
-import { Button } from '@/components/ui/button';
-
 import Period from '@/features/expense/model/types/Period';
 
-interface SelectMonthListItemProps {
-  year: number;
-  month: number;
-  selected: boolean;
-  handleClick: (period: Period) => void;
-}
-
-const SelectMonthListItem: React.FC<SelectMonthListItemProps> = ({
-  year,
-  month,
-  selected,
-  handleClick,
-}) => {
-  return (
-    <li className='w-full h-6'>
-      <Button
-        variant='ghost'
-        className='flex flex-row w-full h-full p-0 font-[15px]'
-        onClick={() => {
-          handleClick({ year, month });
-        }}
-      >
-        <p className='mr-auto'>
-          {year}년 {month}월
-        </p>
-        <div>
-          {selected && (
-            <svg
-              xmlns='http://www.w3.org/2000/svg'
-              width='24'
-              height='24'
-              viewBox='0 0 24 24'
-              fill='none'
-            >
-              <path
-                fill-rule='evenodd'
-                clip-rule='evenodd'
-                d='M21 7.28599L10.4319 17.373C10.2483 17.5483 9.99731 17.6471 9.73541 17.6471C9.4735 17.6471 9.22256 17.5483 9.03891 17.373L3 11.609L4.39299 10.323L9.73541 15.4223L19.607 6L21 7.28599Z'
-                fill='#28A745'
-              />
-            </svg>
-          )}
-        </div>
-      </Button>
-    </li>
-  );
-};
-
+import SelectMonthListItem from './SelectMonthListItem';
 export interface SelectMonthListProps {
   period: Period;
   onSetPeriod: (newPeriod: Period) => void;
